Minor UI updates for existing pages, cleanup and organize
