+++ app/src/main/java/app/jjerrell/meta/course/sample/littlelemon/data/model/UserRegistration.kt
@@ -1,7 +1,7 @@
 package app.jjerrell.meta.course.sample.littlelemon.data.model
 
 data class UserRegistration(
-    val firstName: String,
-    val lastName: String,
-    val email: String
+    val firstName: String = "",
+    val lastName: String = "",
+    val email: String = ""
 )
